+++ cs/DrawTagsClous/Program.cs
@@ -13,12 +13,13 @@ namespace DrawTagsCloud
 		{
             var cloud = new CircularCloudLayouter(new Point(800, 600));
             var rand = new Random();
+            var rectangles = new List<Rectangle>();
             for (var i = 0; i < 1000; i++)
             {
                 var size = new Size(rand.Next(20, 150), rand.Next(20, 150));
-                cloud.PutNextRectangle(size);
+                rectangles.Add(cloud.PutNextRectangle(size));
             }
-            var image = CloudPainter.GetImageOfCloud(cloud, 100);
+            var image = RectanglePainter.GetImageOfRectangles(rectangles, 100);
             var nameImage = "test.jpg";
             image.Save($"..\\..\\..\\{nameImage}");
         }
