+++ cs/TagsCloudVisualization/RectanglePainter.cs
@@ -1,20 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Drawing;
-using System.Text;
+using System.Linq;
 
 namespace TagsCloudVisualization
 {
-    public static class CloudPainter
+    public static class RectanglePainter
     {
-        public static Bitmap GetImageOfCloud(CircularCloudLayouter cloud, int widthOfBorder=0)
+        public static Bitmap GetImageOfRectangles(List<Rectangle> rectangles, int widthOfBorder=0)
         {
-            var cloudBorders = cloud.GetImageSize();
+            var cloudBorders = GetBorderOfRectangles(rectangles);
             var image = new Bitmap(cloudBorders.Width + widthOfBorder * 2, cloudBorders.Height + widthOfBorder * 2);
             var canvas = Graphics.FromImage(image);
             canvas.Clear(Color.White);
             var rand = new Random();
-            var rectangles = cloud.GetRectangles();
             foreach (var rectangle in rectangles)
             {
                 var rect = rectangle;
@@ -25,6 +24,17 @@ namespace TagsCloudVisualization
                 canvas.FillRectangle(new SolidBrush(Color.FromArgb(r, g,b)), rect);
             }
             return image;
+        }
+
+        private static Rectangle GetBorderOfRectangles(List<Rectangle> rectangles)
+        {
+            if (rectangles.Count == 0)
+                throw new ArgumentException("Value cannot be an empty collection.", nameof(rectangles));
+            var maxX = rectangles.Max(rectangle => rectangle.Right);
+            var maxY = rectangles.Max(rectangle => rectangle.Bottom);
+            var minX = rectangles.Min(rectangle => rectangle.Left);
+            var minY = rectangles.Min(rectangle => rectangle.Top);
+            return new Rectangle(minX, minY, maxX - minX, maxY - minY);
         }
     }
 }
